requirements.txt
boto3==1.34.108
boto3==1.38.9
clip==1.0
elevenlabs==1.57.0
mutagen==1.47.0
numpy==2.2.5
openai==1.77.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.7.0.72
Pillow==11.2.1
python-dotenv==1.1.0
Requests==2.32.3
sieve==0.1.9
sievedata==1.4.12
torch==2.0.0
tqdm==4.67.1
tweepy==4.15.0
